lesson4/workspace/requirements.txt
# Lesson 4: Evaluation-Driven Development + Deep Agents
# AI Product Engineer School - Season 3

# Deep Agents SDK
deepagents>=0.3.2

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0

# LangSmith for evaluation
langsmith>=0.1.0

# API clients
httpx>=0.27.0
tavily-python>=0.5.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Data processing
pandas>=2.0.0
